feat(module): add module format support
add module format support
