+++ aktools/__init__.py
@@ -47,8 +47,9 @@ Changelog
 0.0.42: fix: fix email
 0.0.43: add: add support response status code
 0.0.44: add: add login support
+0.0.45: add: add module format support
 """
 
 __title__ = "AKTools"
-__version__ = "0.0.44"
+__version__ = "0.0.45"
 __author__ = "AKFamily"
